product_configurator_search - separate module
