+++ product_configurator_search/__manifest__.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+{
+    'name': 'Product Configurator Search',
+    'version': '11.0',
+    'category': 'Generic Modules/Base',
+    'summary': 'Search product by manufacturer product number',
+    'author': 'Microcom',
+    'license': 'AGPL-3',
+    'website': 'http://www.microcom.ca/',
+    'depends': [
+        'product_configurator_wizard'
+    ],
+    "data": [
+        'wizard/product_configurator_view.xml',
+    ],
+    'demo': [],
+    'test': [],
+    'installable': True,
+    'auto_install': False,
+}
