+++ product_configurator_search/wizard/product_configurator_view.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="product_configurator_form" model="ir.ui.view">
+        <field name="name">product.configurator</field>
+        <field name="model">product.configurator</field>
+        <field name="inherit_id" ref="product_configurator_wizard.product_configurator_form"/>
+        <field name="arch" type="xml">
+            <field name="product_id" position="replace">
+                <field name="search_filter"/>
+                <field name="product_id"/>
+            </field>
+        </field>
+    </record>
+
+</odoo>
